Add files via upload
This is just a simple version of a function calculator. It took me almost a year to finish this project.  I started building it when I had just learnt c++ for about half a year, and this is my first project.  I'm just over 14 years old.  Please don't judge me harshly, but if there's something wrong with my coding, please tell me.  Thank you for your time in reading this.